requirements.txt
streamlit==1.32.0
chromadb==0.4.22
sentence-transformers==2.5.1
google-generativeai==0.8.3
pandas==2.2.0
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1
torch==2.2.0
transformers==4.38.2
numpy==1.26.4
scikit-learn==1.4.0 